oh wow, morelearning is a growing namespace. We got the Intermediate tutorials going on, which will hopefully end up into a small text rpg.
We started a small dungeon adventure story game. really cool idea and very easy implementation, but it is very much a test in stacking classes and heirarchies. def lets you know where variables cannot be placed.

Havent pushed in a few days, hopefully this is fine. I'll be working on the Encounters to the Dungeon after the push.